Fix forms sending post requests when all inputs have not been filled in. Moved ModalFormTitle to shared.
