+++ client/source/components/questionsAnswers/HelpfulQ.jsx
@@ -5,7 +5,7 @@ import ProdContext from '../../ProdContext.js';
 import AnswerForm from './AnswerForm';
 import Modal from '../shared/Modal';
 import useModal from '../../useModal';
-import ModalFormTitle from './ModalFormTitle';
+import ModalFormTitle from '../shared/form/ModalFormTitle';
 
 const HelpfulQ = function CreateHelpfulQComponent({
   questionHelpfulness, id, getQuestions, questionBody
@@ -33,18 +33,23 @@ const HelpfulQ = function CreateHelpfulQComponent({
   };
 
   const handlePost = (event, input) => {
+    let valid = true;
     const keys = Object.keys(input);
     for (let i = 0; i < keys.length; i += 1) {
       if (input[keys[i]] === '') {
+        valid = false;
         alert(`You must enter the following: ${keys[i]}`);
+        break;
       }
     }
-    postAnswer(input)
-      .then((results) => {
-        console.info(results.status);
-        onClose(event);
-      })
-      .catch((err) => console.error('postAnswer error: ', err));
+    if (valid === true) {
+      postAnswer(input)
+        .then((results) => {
+          console.info(results.status);
+          onClose(event);
+        })
+        .catch((err) => console.error('postAnswer error: ', err));
+    }
   };
 
   return (
